Crate wayland_kbd

Keyboard mapping utility for wayland-client using libxkbcommon.

This library provide a simple implemenation for wl_keyboard objects that use libxkbcommon to interpret the keyboard input according to the keymap provided by the compositor.

Usage

To intialize a wl_keyboard with this crate, simply use the provided register_kbd function. See its documentation for details.

Modules

keysyms

Structs

MappedKeyboardImplementation

ModifiersState
Represents the current state of the keyboard modifiers

RMLVO
The RMLVO description of a keymap

Enums

MappedKeyboardError
An error that occured while trying to initialize a mapped keyboard

Functions

register_kbd
Register a keyboard with the implementation provided by this crate

register_kbd_from_rmlvo
Register a keyboard with the implementation provided by this crate