Module wayland_kbd::ffi

Modules

keysyms

Structs

XKBCOMMON_HANDLE

XKBCOMMON_OPTION

XkbCommon

xkb_context

xkb_keymap

xkb_rule_names

xkb_state

xkb_state_component

Enums

xkb_context_flags

xkb_key_direction

xkb_keymap_compile_flags

xkb_keymap_format

xkb_keysym_flags

xkb_log_level

Constants

XKB_KEYCODE_INVALID

XKB_KEYCODE_MAX

XKB_LAYOUT_INVALID

XKB_LED_INVALID

XKB_LED_NAME_CAPS

XKB_LED_NAME_NUM

XKB_LED_NAME_SCROLL

XKB_LEVEL_INVALID

XKB_MOD_INVALID

XKB_MOD_NAME_ALT

XKB_MOD_NAME_CAPS

XKB_MOD_NAME_CTRL

XKB_MOD_NAME_LOGO

XKB_MOD_NAME_NUM

XKB_MOD_NAME_SHIFT

XKB_STATE_LAYOUT_DEPRESSED
Depressed layout, i.e. a key is physically holding it.

XKB_STATE_LAYOUT_EFFECTIVE
Effective layout, i.e. currently active and affects key processing * (derived from the other state components). * Use this unless you explictly care how the state came about.

XKB_STATE_LAYOUT_LATCHED
Latched layout, i.e. will be unset after the next non-modifier * key press.

XKB_STATE_LAYOUT_LOCKED
Locked layout, i.e. will be unset after the key provoking the lock * has been pressed again.

XKB_STATE_LEDS
LEDs (derived from the other state components).

XKB_STATE_MODS_DEPRESSED
Depressed modifiers, i.e. a key is physically holding them.

XKB_STATE_MODS_EFFECTIVE
Effective modifiers, i.e. currently active and affect key * processing (derived from the other state components). * Use this unless you explictly care how the state came about.

XKB_STATE_MODS_LATCHED
Latched modifiers, i.e. will be unset after the next non-modifier * key press.

XKB_STATE_MODS_LOCKED
Locked modifiers, i.e. will be unset after the key provoking the * lock has been pressed again.

Statics

XKBCOMMON_HANDLE

XKBCOMMON_OPTION

Type Definitions

xkb_keycode_t

xkb_keysym_t

xkb_layout_index_t

xkb_layout_mask_t

xkb_led_index_t

xkb_led_mask_t

xkb_level_index_t

xkb_mod_index_t

xkb_mod_mask_t

Implementations