Function wayland_kbd::register_kbd_from_rmlvo

pub fn register_kbd_from_rmlvo<ID: 'static>(
    evqh: &mut EventQueueHandle, 
    kbd: &WlKeyboard, 
    implem: MappedKeyboardImplementation<ID>, 
    idata: ID, 
    rmlvo: RMLVO
) -> Result<(), MappedKeyboardError>

Register a keyboard with the implementation provided by this crate

This requires you to provide an implementation and its implementation data to receive the events after they have been interpreted with the keymap.

The keymap will be loaded from the provided RMLVO rules. Any keymap provided by the compositor will be ignored.

Returns an error if xkbcommon could not be initialized.