Struct wayland_kbd::KbState

pub struct KbState {
    // some fields omitted
}

Methods

impl KbState

fn get_one_sym(&self, keycode: u32) -> u32

Tries to match this keycode as a key symbol according to current keyboard state.

Returns 0 if not possible (meaning that this keycode maps to more than one key symbol).

fn get_utf8(&self, keycode: u32) -> Option<String>

Tries to retrieve the generated keycode as an UTF8 sequence