Struct wayland_kbd::RMLVO

pub struct RMLVO {
    pub rules: Option<String>,
    pub model: Option<String>,
    pub layout: Option<String>,
    pub variant: Option<String>,
    pub options: Option<String>,
}

The RMLVO description of a keymap

All fiels are optional, and the system default will be used if set to None.

Fields
rules: Option<String>

The rules file to use

model: Option<String>

The keyboard model by which to interpret keycodes and LEDs

layout: Option<String>

A comma seperated list of layouts (languages) to include in the keymap

variant: Option<String>

A comma seperated list of variants, one per layout, which may modify or augment the respective layout in various ways

options: Option<String>

A comma seprated list of options, through which the user specifies non-layout related preferences, like which key combinations are used for switching layouts, or which key is the Compose key.