Crate wayland_kbd

Keyboard mapping utility for wayland-client using libxkbcommon.

This library provides a simple wrapper for the wayland Keyboard objects, handling all keymap issues using libxkbcommon in a dynamic way (loading the library dynamically and thus not being linked to it).

To use it, simply call MappedKeyboard::new(..) to wrap you keyboard object and set the key_action callback. This callback will be provided the keycode, the new state of the key (up or down), the keyboard ID, as well as a KbState handle.

This handle will allow you to retrive the keysym associated to the keycode and compare it to the values defined in the keysyms module, or directly restrieve an (utf8) String representation of this character.

Modules

keysyms

Structs

KbState

MappedKeyboard
A wayland keyboard mapped to its keymap