Struct wayland_kbd::MappedKeyboard

pub struct MappedKeyboard<H: Handler> {
    // some fields omitted
}

A wayland keyboard mapped to its keymap

It wraps an event iterator on this keyboard, catching the Keymap, Key, and Modifiers events of the keyboard to handle them using libxkbcommon. All other events are directly forwarded.

Methods

impl<H: Handler> MappedKeyboard<H>

fn new(handler: H) -> Result<MappedKeyboard<H>, MappedKeyboardError>